+++ packages/core/form/src/__tests__/integration/form.js
@@ -16,23 +16,13 @@ const urlFormCreateRepo = getExampleUrl('core', 'form', 'create-repository');
 
 /* Css selectors used for the test */
 const createForm = '[name="create-repo"]';
-const owner = 'div:nth-child(1) > div > [role="status"]';
-const project = 'div:nth-child(2) > div > [role="status"]';
+const owner = 'form > div:nth-child(2) > div > div:nth-child(1)';
+const project = 'form > div:nth-child(2) > div > div:nth-child(2)';
 const repoName = '[name="repo_name"]';
-const accessLevel = 'div:nth-child(4) > div';
-const includeReadme = 'div:nth-child(5) > div > [role="status"]';
+const accessLevel = 'form > div:nth-child(2) > div > div:nth-child(4)';
+const includeReadme = 'form > div:nth-child(2) > div > div:nth-child(5)';
 const createRepoBtn = 'footer > button:nth-child(1)';
 const cancelBtn = 'footer > button:nth-child(2)';
-
-const cssSelectorsForm = [
-  owner,
-  project,
-  repoName,
-  accessLevel,
-  includeReadme,
-  createRepoBtn,
-  cancelBtn,
-];
 
 BrowserTestCase(
   'Create repository form should render without errors',
@@ -41,11 +31,20 @@ BrowserTestCase(
     const formTest = new Page(client);
     await formTest.goto(urlFormCreateRepo);
     await formTest.waitForSelector(createForm);
-
-    cssSelectorsForm.forEach(async cssSelector => {
-      const selectorIsVisible = await formTest.isVisible(cssSelector);
-      expect(selectorIsVisible).toBe(true);
-    });
+    const ownerIsVisible = await formTest.isVisible(owner);
+    const projectIsVisible = await formTest.isVisible(project);
+    const repoNameIsVisible = await formTest.isVisible(repoName);
+    const accessLevelIsVisible = await formTest.isVisible(accessLevel);
+    const includeReadmeIsVisible = await formTest.isVisible(includeReadme);
+    const createRepoBtnIsVisible = await formTest.isVisible(createRepoBtn);
+    const cancelBtnIsVisible = await formTest.isVisible(cancelBtn);
+    expect(ownerIsVisible).toBe(true);
+    expect(projectIsVisible).toBe(true);
+    expect(repoNameIsVisible).toBe(true);
+    expect(accessLevelIsVisible).toBe(true);
+    expect(includeReadmeIsVisible).toBe(true);
+    expect(createRepoBtnIsVisible).toBe(true);
+    expect(cancelBtnIsVisible).toBe(true);
     // TODO: AK-5546: There is an issue with .log
     // if (formTest.log('browser').value) {
     //   formTest.log('browser').value.forEach(val => {