+++ packages/editor/editor-core/src/__tests__/unit/plugins/tasks-and-decisions/ui/ToolbarDecision.tsx
@@ -1,7 +1,11 @@
-import { mount } from 'enzyme';
 import * as React from 'react';
 import { ProviderFactory } from '@atlaskit/editor-common';
-import { doc, p, createEditor } from '@atlaskit/editor-test-helpers';
+import {
+  doc,
+  p,
+  createEditor,
+  mountWithIntl,
+} from '@atlaskit/editor-test-helpers';
 import ToolbarButton from '../../../../../ui/ToolbarButton';
 import ToolbarDecision from '../../../../../plugins/tasks-and-decisions/ui/ToolbarDecision';
 import tasksAndDecisionsPlugin from '../../../../../plugins/tasks-and-decisions';
@@ -20,7 +24,7 @@ describe('@atlaskit/editor-core/ui/ToolbarDecision', () => {
 
   it('should be disabled if isDisabled property is true', () => {
     const { editorView } = editor(doc(p('text')));
-    const toolbarOption = mount(
+    const toolbarOption = mountWithIntl(
       <ToolbarDecision editorView={editorView} isDisabled={true} />,
     );
     expect(toolbarOption.find(ToolbarButton).prop('disabled')).toEqual(true);