+++ packages/editor/editor-core/src/plugins/table/ui/TableFloatingControls/index.tsx
@@ -114,7 +114,6 @@ export default class TableFloatingControls extends Component<Props, State> {
           isTableInDanger={isTableInDanger}
           isHeaderColumnEnabled={isHeaderColumnEnabled}
           isHeaderRowEnabled={isHeaderRowEnabled}
-          isNumberColumnEnabled={isNumberColumnEnabled}
           showInsertButton={showInsertButton}
         />
         <RowControls