+++ packages/editor/renderer/src/text/nodes/applicationCard.ts
@@ -1,8 +1,0 @@
-import { Node as PMNode } from 'prosemirror-model';
-import { NodeReducer } from './';
-
-const applicationCard: NodeReducer = (node: PMNode) => {
-  return `${node.attrs.text || node.attrs.title.text}`;
-};
-
-export default applicationCard;