+++ packages/elements/reactions/examples/examples-util/ReactionsExampleWrapper.tsx
@@ -1,22 +1,43 @@
 import { AnalyticsViewerContainer } from '@atlaskit/analytics-viewer';
 import * as React from 'react';
-import { ReactionClient, ReactionContext } from '../../src';
+import { ReactionClient } from '../../src';
 import { MockReactionsClient } from '../../src/client/MockReactionsClient';
+import {
+  MemoryReactionsStore,
+  ReactionsStore,
+} from '../../src/reaction-store/ReactionsStore';
 
 export type Props = {
   client?: ReactionClient;
-  children: React.ReactChild | React.ReactChild[];
+  children:
+    | ((store: ReactionsStore) => React.ReactChild | React.ReactChild[])
+    | React.ReactChild
+    | React.ReactChild[];
 };
 
-export const ReactionsExampleWrapper: React.StatelessComponent<Props> = ({
-  children,
-  client,
-}: Props) => (
-  <AnalyticsViewerContainer>
-    <ReactionContext client={client}>{children}</ReactionContext>
-  </AnalyticsViewerContainer>
-);
+export class ReactionsExampleWrapper extends React.PureComponent<Props> {
+  static defaultProps = {
+    client: new MockReactionsClient(500),
+  };
+  private store: ReactionsStore;
 
-ReactionsExampleWrapper.defaultProps = {
-  client: new MockReactionsClient(500),
-};
+  constructor(props, context) {
+    super(props, context);
+    this.store = new MemoryReactionsStore(this.props.client!); // default props is handling it
+  }
+
+  renderChildren = () => {
+    if (this.props.children instanceof Function) {
+      return this.props.children(this.store);
+    }
+    return this.props.children;
+  };
+
+  render() {
+    return (
+      <AnalyticsViewerContainer>
+        {this.renderChildren()}
+      </AnalyticsViewerContainer>
+    );
+  }
+}